+++ src/index.js
@@ -4,15 +4,16 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 
 import TempApp from "./pages/TempApp/TempApp";
-import HomePage from "./pages/HomePage";
+import Login from "./pages/Login/Login";
 
 const rootElement = document.getElementById("root");
 ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<HomePage />}>
-          <Route path="TemperatureControl" element={<TempApp />} />
+        <Route path="/">
+          <Route index element={<Login />} />
+          <Route path="/TemperatureControl" element={<TempApp />} />
         </Route>
       </Routes>
     </BrowserRouter>
