+++ src/pages/Login/Loginstyles.css
@@ -1,0 +1,65 @@
+body {
+    font-family: sans-serif;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    min-height: 100vh;
+  }
+  
+  .app-container {
+    height: 896px;
+    width: 414px;
+    background: #2b5870;
+    border-radius: 20px;
+    box-shadow: 10px 10px 38px 0px rgba(0, 0, 0, 0.75);
+  }
+  
+  .start-button-container{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 10%;
+  }
+
+  .start-button {
+    display: flex; 
+    border-radius: 10%;
+    color: #ffffff;
+    height: 50px;
+    width: 120px;
+    align-items:center;
+    border-radius: 0 0 0 0;
+    position: relative;
+    font-size: 48px;
+    border: 3px #ffffff solid;
+    background: #006eff;
+  }
+
+  .start-button  {
+    margin-left: 10px;
+  }
+
+  .temperature-display-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 70%;
+  }
+  
+  .temperature-display {
+    display: flex;
+    border-radius: 50%;
+    color: #ffffff;
+    height: 220px;
+    width: 220px;
+    text-align: center;
+    justify-content: center;
+    align-items: center;
+    font-size: 48px;
+    border: 3px #ffffff solid;
+    transition: background 0.5s;
+  }
+
